structure of the output is now correct
the structure of the raw-sensor data is now correctly formatted. Next the code needs to be cleaned up and tested. The Soilmoisture-sensors are already working though the soilmoisture-sensor itself needs to be calibrated on the receiving side. But the new sht31 is not yet funcitonal and dht has been too unprecise in the past.